refactor: reduce public API for composition validation
