+++ internal/errors/doc.go
